src/main/java/com/example/emazon_microservice_transaction/auth/application/service/JwtTokenValidator.java
package com.example.emazon_microservice_transaction.auth.application.service;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jwts;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

import java.util.Date;

@Component
@RequiredArgsConstructor
public class JwtTokenValidator  {

    private static final String SECRET_KEY = "eENRlFL9i0kFx+JhJGQYk8KGtIwx5Vm/m+aGJXmMNh0=";

    public Claims validateToken(String token) {
        return Jwts.parserBuilder()
                .setSigningKey(SECRET_KEY)
                .build()
                .parseClaimsJws(token)
                .getBody();
    }

    public boolean isTokenExpired(Claims claims) {
        return claims.getExpiration().before(new Date());
    }
}
